project/requirements.txt
nltk
Pillow
pytesseract
regex
requests
tqdm
urllib3
tesseract
